benerin db plus crud event

+++ database/migrations/2024_09_22_054046_create_jemaats_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->text('nama');
             $table->text('jenisKelamin');
             $table->text('alamat');
-            $table->integer('noHp');
+            $table->bigInteger('noHp');
             $table->text('tempatLahir');
             $table->date('tanggalLahir');
             $table->timestamps();
